No default testing channel
