+++ conans/test/remote/auth_test.py
@@ -34,7 +34,7 @@ class AuthorizeTest(unittest.TestCase):
                                                                        ("baduser", "badpass2"),
                                                                        ("pepe", "pepepass")]})
         save(os.path.join(self.conan.current_folder, CONANFILE), conan_content)
-        self.conan.run("export lasote")
+        self.conan.run("export lasote/testing")
         errors = self.conan.run("upload %s" % str(self.conan_reference))
         # Check that return was  ok
         self.assertFalse(errors)
@@ -50,7 +50,7 @@ class AuthorizeTest(unittest.TestCase):
                                                                     ("baduser", "badpass2"),
                                                                     ("baduser3", "badpass3")]})
         save(os.path.join(self.conan.current_folder, CONANFILE), conan_content)
-        self.conan.run("export lasote -p ./ ")
+        self.conan.run("export lasote/testing -p ./ ")
         errors = self.conan.run("upload %s" % str(self.conan_reference), ignore_error=True)
         # Check that return was not ok
         self.assertTrue(errors)
